requirements.txt
dnspython==1.16.0
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
mysql-connector-python==8.0.19
mysql-connector-python-rf==2.2.2
protobuf==3.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
six==1.14.0
